Updated Linux makefile and code. Will compile out of the box on Ubuntu 10.10 with gcc 4.4.5 without warning.

+++ source/game/aotctc_holocron.c
@@ -167,7 +167,7 @@ void AOTCTC_Holocron_Loadpositions( void )
 				}
 			}
 
-			(float)stats[statnum] = (float)atof(t);
+			stats[statnum] = (float)atof(t);
 			statnum++;
 		}
